hw15 is corrected for linters

+++ hw15_go_sql/pkg/pgdb/pgdb.go
@@ -7,13 +7,13 @@ import (
 	"github.com/jackc/pgx/v5/pgxpool"
 )
 
-func New(ctx context.Context, dbDSN string, maxOpenConns int) (*pgxpool.Pool, error) {
+func New(ctx context.Context, dbDSN string, maxOpenConns int32) (*pgxpool.Pool, error) {
 	connConfig, err := pgxpool.ParseConfig(dbDSN)
 	if err != nil {
 		return nil, fmt.Errorf("error creating DSN: %w", err)
 	}
 
-	connConfig.MaxConns = int32(maxOpenConns)
+	connConfig.MaxConns = maxOpenConns
 	connConfig.MinConns = 0
 
 	dbc, err := pgxpool.NewWithConfig(ctx, connConfig)
